+++ app/models/journey.rb
@@ -1,0 +1,4 @@
+class Journey < ApplicationRecord
+  self.implicit_order_column = "created_at"
+  has_many :steps
+end