+++ app/models/plan.rb
@@ -1,4 +1,0 @@
-class Plan < ApplicationRecord
-  self.implicit_order_column = "created_at"
-  has_many :questions
-end