+++ config/environments/test.rb
@@ -59,7 +59,7 @@ Rails.application.configure do
     Bullet.enable = true
     Bullet.bullet_logger = true
     Bullet.raise = true # raise an error if n+1 query occurs
-    Bullet.add_whitelist type: :unused_eager_loading, class_name: "Question", association: :radio_answer
-    Bullet.add_whitelist type: :unused_eager_loading, class_name: "Question", association: :short_text_answer
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: "Step", association: :radio_answer
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: "Step", association: :short_text_answer
   end
 end