+++ spec/factories/journey.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :plan do
+  factory :journey do
     category { "catering" }
     next_entry_id { "47EI2X2T5EDTpJX9WjRR9p" }