+++ spec/factories/step.rb
@@ -1,10 +1,10 @@
 FactoryBot.define do
-  factory :question do
+  factory :step do
     title { "What is your favourite colour?" }
     help_text { "Choose the primary colour closest to your choice" }
     raw { "{\"sys\":{}}" }
 
-    association :plan, factory: :plan
+    association :journey, factory: :journey
 
     trait :radio do
       options { ["Red", "Green", "Blue"] }