+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog 1.0.0].
 - add database foreign key constraints for better data integrity
 - Contentful entry requests use a new Redis read cache
 - service name included in the header
+- refactor name of "Plan" to "Journey"
+- refactor name of "Question" to "Step"
 
 ## [release-002] - 2020-11-16